test(service): AI 분석 기능 단위 테스트 및 로컬 테스트용 API 구현

+++ src/test/java/com/jdc/recipe_service/service/RecipeSearchServiceTest.java
@@ -145,7 +145,7 @@ class RecipeSearchServiceTest {
                 .unit("개")
                 .price(2000)
                 .customName("감자")
-                .customPrice(BigDecimal.valueOf(1000))
+                .customPrice(1000)
                 .customUnit("개")
                 .build();
         when(recipeIngredientRepository.findByRecipeId(existingId))
